feat: Implement granel product handling and WebSocket notifications for new orders
- Added support for granel products in OrdenesPage, allowing users to input a custom amount.
- Integrated WebSocket connection to notify the main PC of new orders for auto-printing.
- Updated PagosPage to reflect current sales and handle payments linked to open cash registers.
- Enhanced ProductosPage to include a checkbox for marking products as granel.
- Modified database schema to accommodate granel products and payment records linked to cash registers.
- Improved ticket printing functionality to display granel items correctly.

+++ POS-Texano/app/models/detalle_orden.py
@@ -17,7 +17,7 @@ class DetalleOrden(Base):
     producto_id: Mapped[int] = mapped_column(
         Integer, ForeignKey("productos.id"), nullable=False
     )
-    cantidad: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
+    cantidad: Mapped[Decimal] = mapped_column(Numeric(10, 3), nullable=False, default=1)
     precio_unitario: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
     subtotal: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
     notas: Mapped[str | None] = mapped_column(String(255), nullable=True)
